refactor : refactor for DataBinding on OnMapErrorActivity
